Commit of the work done
Basic functionalities included + ErrorHandler Class + test file . Xpath is written only, not tested. Will be completed when possible. For the interface a simple while + if case is chosen only for the purpose of the working interface( a pair <string, command> or simply command pattern will be better but there were some complications and time was ticking).